music_video/video_director/video_generator.py
"""
Video Generation Module

Interfaces with Veo3 API to generate video clips from text directives.
"""

import os
from dotenv import load_dotenv

load_dotenv()


class Veo3Generator:
    """
    Client for Veo3 video generation API.
    
    Generates 10-second video clips from text prompts with optional image conditioning.
    """
    
    def __init__(self):
        """Initialize the Veo3 API client."""
        self.api_key = os.getenv("VEO3_API_KEY")
        # TODO: Initialize actual Veo3 client when API details are available
    
    def generate_video(
        self,
        prompt: str,
        duration: int = 10,
        reference_image_path: str = None,
        output_path: str = None
    ) -> str:
        """
        Generate a video clip from a text prompt.
        
        Args:
            prompt: Text directive describing the desired video
            duration: Video duration in seconds (default: 10)
            reference_image_path: Optional image for visual continuity
            output_path: Where to save the generated video
            
        Returns:
            Path to the generated video file
            
        Example:
            generator = Veo3Generator()
            video_path = generator.generate_video(
                prompt="A dancer moves gracefully through a neon-lit cityscape...",
                reference_image_path="previous_frame.jpg",
                output_path="segment_1.mp4"
            )
        """
        # TODO: Implement this function
        # Steps:
        # 1. Prepare the API request with prompt and optional reference image
        # 2. Call Veo3 API
        # 3. Download/save the generated video
        # 4. Return the output path
        pass
