music_video/requirements.txt
openai
python-dotenv
pydub
opencv-python
moviepy
requests
numpy
